+++ shared/domain/shows/api/src/commonTest/kotlin/com/thomaskioko/tvmaniac/shows/api/MockData.kt
@@ -58,7 +58,7 @@ val errorShowResult = ShowResult(
     )
 )
 
-fun categoryResult(categoryId: Int) = listOf(
+fun categoryResult(categoryId: Long) = listOf(
     SelectShowsByCategory(
         trakt_id = 84958,
         tmdb_id = 849583,
